Backend/requirements.txt
fastapi
uvicorn
pymongo
numpy
pandas
transformers
scikit-learn
faiss-cpu
sentencepiece
torch
pydantic
python-dotenv